Move NormalizeCSVCommand to Tool namespace
